+++ src/core/loss_detection.c
@@ -378,9 +378,9 @@ QuicLossDetectionUpdateTimer(
     } else {
         QuicTraceEvent(
             ConnLossDetectionTimerSet,
-            "[conn][%p] Setting loss detection %d timer for %u us. (ProbeCount=%hu)",
+            "[conn][%p] Setting loss detection %hhu timer for %u us. (ProbeCount=%hu)",
             Connection,
-            TimeoutType,
+            (uint8_t)TimeoutType,
             Delay,
             LossDetection->ProbeCount);
         UNREFERENCED_PARAMETER(TimeoutType);
@@ -412,9 +412,9 @@ QuicLossDetectionOnPacketSent(
         //
         QuicTraceEvent(
             AllocFailure,
-            "Allocation of '%s' failed. (%zu bytes)",
+            "Allocation of '%s' failed. (%llu bytes)",
             "Sent packet metadata",
-            SIZEOF_QUIC_SENT_PACKET_METADATA(TempSentPacket->FrameCount));
+            (uint64_t)SIZEOF_QUIC_SENT_PACKET_METADATA(TempSentPacket->FrameCount));
         QuicLossDetectionRetransmitFrames(LossDetection, TempSentPacket, FALSE);
         QuicSentPacketMetadataReleaseFrames(TempSentPacket);
         return;
@@ -996,11 +996,11 @@ QuicLossDetectionDetectAndHandleLostPackets(
                         LossDetection->LargestAck - Packet->PacketNumber);
                     QuicTraceEvent(
                         ConnPacketLost,
-                        "[conn][%p][TX][%llu] %hhu Lost: %d",
+                        "[conn][%p][TX][%llu] %hhu Lost: %hhu",
                         Connection,
                         Packet->PacketNumber,
                         QuicPacketTraceType(Packet),
-                        QUIC_TRACE_PACKET_LOSS_FACK);
+                        (uint8_t)QUIC_TRACE_PACKET_LOSS_FACK);
                 }
             } else if (Packet->PacketNumber < LossDetection->LargestAck &&
                         CxPlatTimeAtOrBefore32(Packet->SentTime + TimeReorderThreshold, TimeNow)) {
@@ -1013,11 +1013,11 @@ QuicLossDetectionDetectAndHandleLostPackets(
                         CxPlatTimeDiff32(Packet->SentTime, TimeNow));
                     QuicTraceEvent(
                         ConnPacketLost,
-                        "[conn][%p][TX][%llu] %hhu Lost: %d",
+                        "[conn][%p][TX][%llu] %hhu Lost: %hhu",
                         Connection,
                         Packet->PacketNumber,
                         QuicPacketTraceType(Packet),
-                        QUIC_TRACE_PACKET_LOSS_RACK);
+                        (uint8_t)QUIC_TRACE_PACKET_LOSS_RACK);
                 }
             } else {
                 break;
@@ -1682,11 +1682,11 @@ QuicLossDetectionScheduleProbe(
                 Packet->PacketNumber);
             QuicTraceEvent(
                 ConnPacketLost,
-                "[conn][%p][TX][%llu] %hhu Lost: %d",
+                "[conn][%p][TX][%llu] %hhu Lost: %hhu",
                 Connection,
                 Packet->PacketNumber,
                 QuicPacketTraceType(Packet),
-                QUIC_TRACE_PACKET_LOSS_PROBE);
+                (uint8_t)QUIC_TRACE_PACKET_LOSS_PROBE);
             if (QuicLossDetectionRetransmitFrames(LossDetection, Packet, FALSE) &&
                 --NumPackets == 0) {
                 return;
